Create BootupPowershell-CorrectAccessPoint.bat
This file will be neccesary to be ran before any of the tools can be used if inside a "secured" environment.
Anything you do, do it legally. No harm done by anyone can be blamed on me. All tools should be used to convenience. 
Any Issues, please report it to me.